Start on version control
This appears to be a start on the Version Control & related pages
(Issue #38).

IIRC (it's been a while since I started this file), the goal was to
document what VCS is, what git, Github, & Markdown are, as supporting
documents for a "How to Contribute to StepZero" doc so I could start
getting other people to lend a hand.

So I guess I should probably get back to that sometime RSN.